Move fast_spec_helper.rb to spec root
