+++ spec/specialist_document_header_extractor_spec.rb
@@ -1,4 +1,4 @@
-require "support/fast_spec_helper"
+require "fast_spec_helper"
 
 require "specialist_document_header_extractor"
